azure-cloud-resume-challenge/backend/backend/api/requirements.txt
azure-functions
azure-data-tables
azure-cosmos